+++ src/main/kotlin/com/github/strindberg/emacsj/search/CommonHighlighter.kt
@@ -19,7 +19,7 @@ object CommonHighlighter {
     @VisibleForTesting
     internal var testing = false
 
-    private var progressIndicators = mutableListOf<ProgressIndicator>()
+    private val progressIndicators = mutableListOf<ProgressIndicator>()
 
     internal fun cancel(editor: Editor) {
         val iterator = progressIndicators.iterator()